Move some bitstring-related functions from mathtools to a new header
