+++ test/CMakeLists.txt
@@ -58,7 +58,8 @@ endfunction()
 
 add_test_target(TARGET circuit_test SOURCES "source/circuit_test.cpp")
 add_test_target(TARGET gate_test SOURCES "source/gate_test.cpp")
-add_test_target(TARGET state_test SOURCES "source/state_test.cpp")
+add_test_target(TARGET state_test SOURCES "source/state/state_test.cpp")
+add_test_target(TARGET bitstring_utils_test SOURCES "source/state/bitstring_utils_test.cpp")
 add_test_target(TARGET operations_test SOURCES "source/operations_test.cpp")
 add_test_target(TARGET gate_pair_generator_test SOURCES "source/gate_pair_generator_test.cpp")
 add_test_target(TARGET simulate_test SOURCES "source/simulate_test.cpp")
